Move domain search tracking functions to its own hook
